[PCS-19] Advertisement component
