+++ frontend/src/components/Advertisement/index.js
@@ -1,29 +1,47 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Grid, Paper, Rating } from '@mui/material';
-import { Box } from '@mui/system';
+import FmdGoodIcon from '@mui/icons-material/FmdGood';
+import { Paper, Rating, Box, Typography } from '@mui/material';
+import Icon from '../Icon';
 import PriceInterval from '../PriceInterval';
-import Tag from '../Tag';
 import styles from './styles';
+import renderTags from './utils';
 
-const Advertisement = ({ activities, animals, starsValue, price }) => {
+const Advertisement = ({
+  activities,
+  animals,
+  starsValue,
+  price,
+  location,
+  image,
+}) => {
   return (
     <Paper sx={styles.root}>
       <Box p={10}>
-        <Box display={'flex'} sx={styles.root}>
-          <Box flexGrow={3}>{'img'}</Box>
+        <Box display={'flex'} sx={styles.root} gap={20}>
+          <Box
+            component="img"
+            sx={{
+              height: 140,
+              borderRadius: 25,
+            }}
+            src={image}
+          />
           <Box
             display={'flex'}
-            flexGrow={9}
+            flexGrow={10}
             flexDirection={'column'}
             justifyContent={'space-between'}
           >
             <Box display={'flex'} flexDirection={'column'}>
-              <Box>{renderTags(activities, 'activitiesTags', 2)}</Box>
-              <Box>{renderTags(animals, 'animalsTags', 2)}</Box>
+              {renderTags(activities, 'activitiesTags', 2)}
+              {renderTags(animals, 'animalsTags', 2)}
             </Box>
             <Box display={'flex'} flexDirection={'column'}>
-              <Box display={'flex'}>{'tags3'}</Box>
+              <Box display={'flex'}>
+                <Icon Component={FmdGoodIcon} size={'big'}></Icon>
+                <Typography variant={'h1'}>{location}</Typography>
+              </Box>
             </Box>
           </Box>
           <Box
@@ -49,29 +67,15 @@ const Advertisement = ({ activities, animals, starsValue, price }) => {
   );
 };
 
-const renderTags = (tagLabels, styleType, amountToFit) => {
-  const size = tagLabels.length;
-  const activitiesTags = [];
-  const color = styles[styleType].backgroundColor;
-  for (let i = 0; i < amountToFit; i++) {
-    if (size > i) {
-      activitiesTags.push(<Tag label={tagLabels[i]} color={color} />);
-    }
-  }
-  if (size > amountToFit) {
-    activitiesTags.push(
-      <Tag label={'+' + (size - amountToFit)} color={color} />
-    );
-  }
-
-  return activitiesTags;
-};
-
 Advertisement.propTypes = {
   activities: PropTypes.array.isRequired,
   animals: PropTypes.array.isRequired,
   starsValue: PropTypes.number.isRequired,
   price: PropTypes.object.isRequired,
+  location: PropTypes.string.isRequired,
+  image: PropTypes.string.isRequired,
 };
+
+//TODO Location prop as an object
 
 export default Advertisement;
