+++ frontend/src/components/Advertisement/stories.js
@@ -24,6 +24,9 @@ Advertisement.args = {
     to: 20,
     priceType: 'hourly',
   },
+  location: 'Kraków',
+  image:
+    'https://media.istockphoto.com/photos/young-redhead-woman-hug-her-small-mixedbreed-dog-picture-id485251750?b=1&k=20&m=485251750&s=170667a&w=0&h=v7Wf4tPLnkGl_yZxXiWyjzKYdLgirM_zieYBLilAM5c=',
 };
 
 export { Advertisement };
