+++ frontend/src/components/Advertisement/styles.js
@@ -6,15 +6,8 @@ export default {
     backgroundColor: (theme) => theme.palette.neutral.main,
   },
   root: {
-    // minWidth: 120,
-    minHeight: 120,
-    borderRadius: 3,
-    // color: (theme) => theme.palette[color].contrastText,
-    // textTransform: 'none',
-    // fontSize: (theme) => theme.fontSize.big,
-    // backgroundColor: (theme) => theme.palette[color].main,
-    // '&:hover': {
-    //   backgroundColor: (theme) => theme.palette[color].dark,
-    // },
+    minHeight: 140,
+    borderRadius: 25,
+    p: 5,
   },
 };
